Update input screen.py
Can now write the objects on the input menu to a file. The code uses the pickle library to write the objects in a byte form so that they can later be read back as a frog object. (I haven't tried to read it back off the file but we should be able to add that to the main code relatively easily.) When using this you must press the 'close file' button before closing the window as this may corrupt the file.